+++ tests/helper_utils.py
@@ -13,7 +13,7 @@ import benchbuild.source.base as base
 import benchbuild.utils.settings as bb_settings
 import plumbum as pb
 from benchbuild import Project
-from benchbuild.source import Git
+from benchbuild.source import Git, FetchableSource, Variant
 from benchbuild.utils.cmd import git
 
 import varats.utils.settings as settings
@@ -283,3 +283,35 @@ class ConfigurationHelper:
         )
         test_config.add_config_option(ConfigurationOptionImpl("buzz", "None"))
         return test_config
+
+
+class BBTestSource(FetchableSource):
+    """Source test fixture class."""
+
+    test_versions: tp.List[str]
+
+    def __init__(
+        self, test_versions: tp.List[str], local: str,
+        remote: tp.Union[str, tp.Dict[str, str]]
+    ):
+        super().__init__(local, remote)
+        self.test_versions = test_versions
+
+    @property
+    def local(self) -> str:
+        return "test_source"
+
+    @property
+    def remote(self) -> tp.Union[str, tp.Dict[str, str]]:
+        return "test_remote"
+
+    @property
+    def default(self) -> Variant:
+        return Variant(owner=self, version=self.test_versions[0])
+
+    # pylint: disable=unused-argument,no-self-use
+    def version(self, target_dir: str, version: str) -> pb.LocalPath:
+        return pb.local.path('.') / f'varats-test-{version}'
+
+    def versions(self) -> tp.Iterable[Variant]:
+        return [Variant(self, v) for v in self.test_versions]
