+++ tests/mapping/test_configuration_map.py
@@ -5,7 +5,7 @@ import unittest.mock as mock
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 
-from tests.test_utils import ConfigurationHelper
+from tests.helper_utils import ConfigurationHelper
 from varats.base.configuration import (
     DummyConfiguration,
     ConfigurationImpl,
