+++ tests/table/test_code_centrality_table.py
@@ -1,7 +1,7 @@
 """Test bug overview table."""
 import unittest
 
-from tests.test_utils import run_in_test_environment, UnitTestFixtures
+from tests.helper_utils import run_in_test_environment, UnitTestFixtures
 from varats.paper.paper_config import load_paper_config, get_loaded_paper_config
 from varats.projects.discover_projects import initialize_projects
 from varats.table.tables import TableConfig, TableFormat
