+++ tests/tools/test_driver_artefacts.py
@@ -4,7 +4,7 @@ import unittest
 
 from click.testing import CliRunner
 
-from tests.test_utils import run_in_test_environment, UnitTestFixtures
+from tests.helper_utils import run_in_test_environment, UnitTestFixtures
 from varats.data.discover_reports import initialize_reports
 from varats.paper.paper_config import get_paper_config, load_paper_config
 from varats.paper_mgmt.artefacts import Artefact, load_artefacts
