+++ tests/tools/test_driver_buildsetup.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from click.testing import CliRunner
 
-from tests.test_utils import run_in_test_environment
+from tests.helper_utils import run_in_test_environment
 from varats.containers.containers import (
     ImageBase,
     get_image_name,
