+++ tests/tools/test_driver_casestudy.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from click.testing import CliRunner
 
-from tests.test_utils import (
+from tests.helper_utils import (
     run_in_test_environment,
     TEST_INPUTS_DIR,
     UnitTestFixtures,
